requirements.txt
PyQt6
firebase-admin
google-cloud-firestore
reportlab
firebase 
firebase-auth
pyinstaller